refactor(pci-rancher): update api and hook folder structure ref: TAPC-987
