--- packages/manager/apps/pci-rancher/src/api/index.ts
@@ -1,1 +1,0 @@
-export * from './apiv2/services';
