+++ packages/manager/apps/pci-rancher/src/components/Modal/DeleteModal/DeleteModal.component.tsx
@@ -18,8 +18,11 @@ import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { RancherService } from '@/api/api.type';
-import { useTrackingAction, useTrackingPage } from '@/hooks/useTrackingPage';
+import { RancherService } from '@/types/api.type';
+import {
+  useTrackingAction,
+  useTrackingPage,
+} from '@/hooks/useTrackingPage/useTrackingPage';
 import { TrackingEvent, TrackingPageView } from '@/utils/tracking';
 import Modal from '../Modal.component';
