+++ packages/manager/apps/pci-rancher/src/components/StatusChip/StatusChip.component.tsx
@@ -3,7 +3,7 @@ import { ODS_SPINNER_SIZE } from '@ovhcloud/ods-components';
 import { OsdsChip, OsdsSpinner } from '@ovhcloud/ods-components/react';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { ResourceStatus } from '@/api/api.type';
+import { ResourceStatus } from '@/types/api.type';
 
 const StatusChip = ({ label }: { label: string }) => {
   const { t } = useTranslation('pci-rancher/listing');
