+++ packages/manager/apps/pci-rancher/src/components/Table/ActionsCell/ActionsCell.test.tsx
@@ -3,7 +3,7 @@ import { Row } from '@tanstack/react-table';
 import ActionsCell from './ActionsCell.component';
 import { render, waitFor } from '../../../utils/test/test.provider';
 import { rancherError, rancherMocked } from '../../../_mock_/rancher';
-import { RancherService } from '@/api/api.type';
+import { RancherService } from '@/types/api.type';
 import listingTranslation from '../../../public/translations/pci-rancher/listing/Messages_fr_FR.json';
 
 const setupSpecTest = async (rancherService: RancherService = rancherMocked) =>
