+++ packages/manager/apps/pci-rancher/src/components/UpdateRancherVersionBanner/UpdateVersionBanner.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, waitFor } from '@/utils/test/test.provider';
 import { rancherMocked } from '@/_mock_/rancher';
 import { versionsMocked } from '@/_mock_/version';
-import { ResourceStatus } from '@/api/api.type';
+import { ResourceStatus } from '@/types/api.type';
 import UpdateVersionBanner from './UpdateVersionBanner.component';
 import updateSoftwareTranslation from '@/public/translations/pci-rancher/updateSoftware/Messages_fr_FR.json';
