+++ packages/manager/apps/pci-rancher/src/components/layout-helpers/OdsCell/ProductStatusCell.component.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { TFunction } from 'i18next';
 import { ODS_SPINNER_SIZE } from '@ovhcloud/ods-components';
 import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
-import { ResourceStatus } from '@/api/api.type';
+import { ResourceStatus } from '@/types/api.type';
 
 // TODO: remove this file
 export type DataGridCellProps<Cell = string> = {
