+++ packages/manager/apps/pci-rancher/src/pages/listing/delete/DeleteRancherModal.tsx
@@ -2,7 +2,10 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import Loading from '@/components/Loading/Loading.component';
 import DeleteModal from '@/components/Modal/DeleteModal/DeleteModal.component';
-import { useDeleteRancher, useRancher } from '@/hooks/useRancher';
+import {
+  useDeleteRancher,
+  useRancher,
+} from '@/data/hooks/useRancher/useRancher';
 
 const DeleteRancherModal = () => {
   const { data: selectedRancher, isLoading } = useRancher();
