+++ packages/manager/apps/pci-rancher/src/pages/listing/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Listing, { ListingProps } from './index';
 import { render, waitFor } from '../../utils/test/test.provider';
-import { RancherService } from '../../api/api.type';
+import { RancherService } from '../../types/api.type';
 import listingTranslation from '../../public/translations/pci-rancher/listing/Messages_fr_FR.json';
 import { rancherMocked } from '../../_mock_/rancher';
