+++ packages/manager/apps/pci-rancher/src/pages/listing/index.tsx
@@ -15,12 +15,14 @@ import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useHref, useNavigate, useParams } from 'react-router-dom';
 import { Title } from '@ovhcloud/manager-components';
-import { RancherService } from '@/api/api.type';
+import { RancherService } from '@/types/api.type';
 import TableContainer from '@/components/Table/TableContainer/TableContainer.component';
-import { useTrackingAction } from '@/hooks/useTrackingPage';
+import {
+  useTrackingAction,
+  useTrackingPage,
+} from '@/hooks/useTrackingPage/useTrackingPage';
 import { getOnboardingUrl } from '@/utils/route';
 import { TrackingEvent, TrackingPageView } from '@/utils/tracking';
-import { useTrackingPage } from '../../hooks/useTrackingPage';
 import RancherTaskMessage from './RancherTaskMessage';
 
 export interface ListingProps {
