+++ packages/manager/apps/pci-rancher/src/pages/onboarding/index.tsx
@@ -9,9 +9,12 @@ import { useNavigate, useParams } from 'react-router-dom';
 import onboardingImgSrc from '../../assets/onboarding-img.png';
 import Breadcrumb from '../../components/Breadcrumb/Breadcrumb.component';
 import { getCreateRancherUrl } from '@/utils/route';
-import { useTrackingAction, useTrackingPage } from '@/hooks/useTrackingPage';
+import {
+  useTrackingAction,
+  useTrackingPage,
+} from '@/hooks/useTrackingPage/useTrackingPage';
 import { TrackingEvent, TrackingPageView } from '@/utils/tracking';
-import { useGuideUtils } from '@/hooks/guideLink/useGuideLink';
+import { useGuideUtils } from '@/hooks/useGuideLink/useGuideLink';
 
 export default function Onboarding() {
   const { t } = useTranslation('pci-rancher/onboarding');
